Complete review of unit & integration tests

+++ vtl-api/src/main/java/it/bancaditalia/oss/vtl/model/transform/analytic/LimitCriterion.java
@@ -40,5 +40,5 @@ public interface LimitCriterion
 	/**
 	 * @return The distance of the bound from current data point
 	 */
-	public long getCount();
+	public int getCount();
 }
