+++ vtl-transform/src/main/java/it/bancaditalia/oss/vtl/impl/transform/aggregation/OffsetTransformation.java
@@ -50,8 +50,8 @@ public class OffsetTransformation extends SimpleAnalyticTransformation
 	{
 		super(FIRST_VALUE, operand, partitionBy, orderByClause, 
 				new WindowCriterionImpl(DATAPOINTS, 
-						new LimitClause(direction == LAG ? PRECEDING : FOLLOWING, Math.abs(offset.get())),
-						new LimitClause(direction == LAG ? PRECEDING : FOLLOWING, Math.abs(offset.get()))));
+						new LimitClause(direction == LAG ? PRECEDING : FOLLOWING, Math.abs(offset.get().intValue())),
+						new LimitClause(direction == LAG ? PRECEDING : FOLLOWING, Math.abs(offset.get().intValue()))));
 
 		this.defaultValue = defaultValue;
 	}
